Simplify docs test check
